files paths refactoring
