+++ src/components/index.js
@@ -2,9 +2,12 @@ export { EditTestsFilter } from './EditTests/EditTestsFilter/EditTestsFilter';
 export { EditTestsTable } from './EditTests/EditTestsTable/EditTestsTable';
 export { TestLevelSelector } from './TestLevelSelector/TestLevelSelector';
 export { Essay } from './Test/Essay/Essay';
+export { EssayReportMistake } from './Test/Essay/EssayReportMistake';
 export { Grammar } from './Test/Grammar/Grammar';
 export { Listening } from './Test/Listening/Listening';
+export { ListeningReportMistake } from './Test/Listening/ListeningReportMistake';
 export { Speaking } from './Test/Speaking/Speaking';
+export { SpeakingReportMistake } from './Test/Speaking/SpeakingReportMistake';
 export { Player } from './Test/Player/Player';
 export { listeningTasks } from './Test/testData/ListeningTasks';
 export { grammarTasks } from './Test/testData/GrammarTasks';
