+++ src/pages/Test/IconClose.scss
